+++ project2026/app/Http/Controllers/AttendanceControllerHCNS.php
@@ -13,12 +13,20 @@ class AttendanceControllerHCNS extends Controller
 
     public function index()
     {
+        if (auth()->user()->role->name !== 'hcns') 
+        {
+            return back();
+        }
         $attendances = Attendance::with('user')->orderBy('work_date', 'desc')->where('confirm','yes')->get();
         return view('hcns.attendances.index', compact('attendances'));
     }
 
     public function edit($id)
     {
+        if (auth()->user()->role->name !== 'hcns') 
+        {
+            return back();
+        }
         $attendance = Attendance::findOrFail($id);
         $employees = Employee::all();
         return view('hcns.attendances.edit', compact('attendance', 'employees'));
@@ -26,6 +34,11 @@ class AttendanceControllerHCNS extends Controller
 
     public function update(Request $request, $id)
     {
+        if (auth()->user()->role->name !== 'hcns') 
+        {
+            return back();
+        }
+
         $request->validate([
             'work_date' => 'required|date'
         ],[
@@ -59,6 +72,10 @@ class AttendanceControllerHCNS extends Controller
 
     public function delete($id)
     {
+        if (auth()->user()->role->name !== 'hcns') 
+        {
+            return back();
+        }
         Attendance::findOrFail($id)->delete();
         return back()->with('success','Xóa dữ liệu chấm công thành công');
     }
